Fix tejas's pic to his preferred fallback while he
looks for a picture that he is most comfortable putting up on our
website.

cf. his reply via email to 2020@inclojure.org
sub: Talk Selection for IN/Clojure 2020

Yes, feel free to name me and use links on website and promo media.

I'll share a photo and bio soon. (or give me options if you have a photo
of me? I am not too happy with either the one on Quintype or Nilenso
website). Or fallback on this one
https://imgee.s3.amazonaws.com/imgee/ec7d04942fb0467791190934936a49cd.jpeg