Emergency fix CFP section with date & clarity on the
two-phase talk selection. Also link to the tentative schedule at the
hasgeek portal as a quick fix for the problem of schedule invisibility.

twitter.com/amarjeet000 reported disappointment at the fact that the
deadline was not obvious, and that the two-phase selection appeared
shady, because, again, it was not stated clearly in our communication.

HOW did we miss publicising these facts prominently? :facepalm:

I don't have time for a proper solution right now to publish the
schedule properly and/or list selected talks.

We could fetch the JSON source from the hasgeek site, and embed it
in our site, so we have one source of truth.